+++ probables/__init__.py
@@ -3,7 +3,7 @@ from __future__ import (unicode_literals, absolute_import, print_function)
 from .blooms import (BloomFilter, BloomFilterOnDisk, CountingBloomFilter)
 from .countminsketch import (CountMinSketch, HeavyHitters, StreamThreshold,
                              CountMeanSketch, CountMeanMinSketch)
-from .cuckoo import (CuckooFilter)
+from .cuckoo import (CuckooFilter, ExpandingCuckooFilter)
 from .exceptions import (InitializationError, NotSupportedError,
                          ProbablesBaseException, CuckooFilterFullError)
 
@@ -19,4 +19,4 @@ __bugtrack_url__ = 'https://github.com/barrust/pyprobables/issues'
 __all__ = ['BloomFilter', 'BloomFilterOnDisk', 'CountMinSketch',
            'HeavyHitters', 'StreamThreshold', 'CountMeanSketch',
            'CountMeanMinSketch', 'CountingBloomFilter', 'CuckooFilter',
-           'CuckooFilterFullError']
+           'CuckooFilterFullError', 'ExpandingCuckooFilter']