+++ probables/cuckoo/__init__.py
@@ -1,3 +1,3 @@
 ''' Cuckoo Filter data structures '''
 from __future__ import (unicode_literals, absolute_import, print_function)
-from . cuckoo import (CuckooFilter)
+from . cuckoo import (CuckooFilter, ExpandingCuckooFilter)