+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * CuckooFilter
     * Added Import / Export functionality
     * Enforce single insertion per key
+    * ExpandingCuckooFilter - auto expands when insertion failure OR
+    when called to do so
 
 ### Version 0.1.0:
 * Cuckoo Filter